requirements.txt
flask
tensorflow
pillow
numpy
gunicorn
datetime
google-cloud-firestore
google-cloud-storage